Add a "the" before "ists"
I got this output:

"Joe Rogan is a member of feminists. The Trump administration were originally opposed to this but they're now in on it. I know because feminists have included it in their manifesto, and if you Google for feminists there's loads of info."

This just changes the first instance of the name to "the" for
"feminists", "Satanists", "Socialists" and "atheists".